+++ test/mpi/rma/putpscw1.c
@@ -42,7 +42,7 @@ int main( int argc, char *argv[] )
 
 		MPI_Type_extent( recvtype.datatype, &extent );
 		MPI_Win_create( recvtype.buf, recvtype.count * extent, 
-				extent, MPI_INFO_NULL, comm, &win );
+				(int)extent, MPI_INFO_NULL, comm, &win );
 		MPI_Win_get_group( win, &wingroup );
 		if (rank == source) {
 		    /* To improve reporting of problems about operations, we
